MainWindow: Re-display ConnectionDialog if an error occurred
This required actually quite tricky exception handling because
we can't just close the dialog since MainWindow then expects
the settings to be written at their location. Also, we want
to retry every time after a failure but still give the user the
possiblity to actually abort to establish the connection.
And since this all happens in constructors, it really got dodgy.

Closes #30